changed return and rent logic

+++ customers/models.py
@@ -18,7 +18,7 @@ class Customodel(models.Model):
     phone=models.CharField(max_length=15)
     favorite_station_id=models.IntegerField(null=True)
     active_trip_id=models.IntegerField(null=True)
-    credits=models.IntegerField(default=10)
+    credits=models.FloatField(default=10.00)
     session_id=models.IntegerField(null=True)
 
     class Meta:
